refactor: StoreMenus, StoreMenusImpl 클래스 생성
